[Front] fix(skills): correct topViews/nonTopViews filter logic
