+++ front/components/agent_builder/capabilities/capabilities_sheet/types.ts
@@ -1,28 +1,10 @@
 import type { AgentBuilderSkillsType } from "@app/components/agent_builder/AgentBuilderFormContext";
+import type { SelectedTool } from "@app/components/agent_builder/capabilities/shared/types";
 import type {
   CapabilitiesSheetState,
   SheetState,
 } from "@app/components/agent_builder/skills/types";
-import type { MCPServerViewTypeWithLabel } from "@app/components/shared/tools_picker/MCPServerViewsContext";
 import type { BuilderAction } from "@app/components/shared/tools_picker/types";
-import { AGENT_MEMORY_SERVER_NAME } from "@app/lib/actions/mcp_internal_actions/constants";
-
-// TODO(skills 2025-12-18): duplicated from MCPServerViewsSheet, to cleanup later
-export const TOP_MCP_SERVER_VIEWS = [
-  "web_search_&_browse",
-  "image_generation",
-  AGENT_MEMORY_SERVER_NAME,
-  "deep_dive",
-  "interactive_content",
-  "slack",
-  "gmail",
-  "google_calendar",
-];
-
-export type SelectedTool = {
-  view: MCPServerViewTypeWithLabel;
-  configuredAction?: BuilderAction;
-};
 
 export type CapabilityFilterType = "all" | "tools" | "skills";
