+++ front/components/agent_builder/capabilities/mcp/utils/actionNameUtils.ts
@@ -1,4 +1,4 @@
-import type { SelectedTool } from "@app/components/agent_builder/capabilities/mcp/MCPServerViewsSheet";
+import type { SelectedTool } from "@app/components/agent_builder/capabilities/shared/types";
 import type { BuilderAction } from "@app/components/shared/tools_picker/types";
 
 interface GenerateUniqueActionNameParams {
