+++ front/components/agent_builder/capabilities/shared/types.ts
@@ -1,0 +1,19 @@
+import type { MCPServerViewTypeWithLabel } from "@app/components/shared/tools_picker/MCPServerViewsContext";
+import type { BuilderAction } from "@app/components/shared/tools_picker/types";
+import { AGENT_MEMORY_SERVER_NAME } from "@app/lib/actions/mcp_internal_actions/constants";
+
+export type SelectedTool = {
+  view: MCPServerViewTypeWithLabel;
+  configuredAction?: BuilderAction;
+};
+
+export const TOP_MCP_SERVER_VIEWS = [
+  "web_search_&_browse",
+  "image_generation",
+  AGENT_MEMORY_SERVER_NAME,
+  "deep_dive",
+  "interactive_content",
+  "slack",
+  "gmail",
+  "google_calendar",
+];
